- Reviewing hamming distance, I realized how it was supposed to be working and fixed it.  challenge6 doesn't feel so hacky anymore

+++ Utils.h
@@ -22,7 +22,7 @@ namespace cipher {
     std::string fixed_xor(const std::string& bytes1, const std::string& bytes2);
     std::string byte_xor(const std::string& bytes, unsigned char byte);
     std::string repeating_xor(const std::string& bytes, const std::string& repeat);
-    size_t hamming_distance(const std::string& bytes1, const std::string& bytes2);
+    double hamming_distance(const std::string& bytes1, const std::string& bytes2);
     unsigned char guess_xor_byte(const std::string& bytes, float* out_score = nullptr);
     std::string pad_pkcs7(const std::string& bytes, unsigned char block_size);
     std::string unpad_pkcs7(const std::string& bytes, unsigned char block_size);
